# Simple decomposing of interpolation for weights re-use.