This is pretty much the extent of the changes I propose for seeing the following performance boost when regridding. My back of an envelope notes on performance:

Source grid size | Target grid size | Before (seconds) | After (seconds)
---------------------|---------------------|------------------------|--------------------
1000 x 1001 | 800 x 999 | 0.76 | 0.76
1000 x 1001 x 8 | 800 x 999 | 5.7 | 4.1
1000 x 1001 x 80 | 800 x 999 | 53 | 43

Clearly there is still a lot of room for improvement though.