@rhattersley 2015-05-19
> 53 -> 43

:+1: :grin: 

Any thoughts on pushing this upstream?