@rhattersley 2015-05-19
Summary of offline conversation with @pelson:
 - Let's leave this unmerged for now and accumulate more changes on this branch.
 - When we know enough about what we want to do (e.g. existing vs. sparse matrix vs ???) we can consider pushing upstream, and then decide what to do within Iris.