@rhattersley 2015-05-20
I've not thought this through completely, but we might want to re-consider the handling of pre-computed "weights" (we could do with a more general term) for "linear" vs. "nearest".