@rhattersley 2015-05-20
I've combined this with a proof-of-concept sparse matrix refactor. The sparse matrix version is faster even when not re-using the prepared weights, but the combination really flies.

Source grid | Target grid | Prepared / s | Prepared and sparse / s | Speed up
---------------------|---------------------|------------------------|--------------------|------
1152 x 1536 | 1200 x 1600 | 1.1 | 1.0 | 1.1
30 x 1152 x 1536 | 1200 x 1600 | 15.6 | 2.4 | 6.5
70 x 1152 x 1536 | 1200 x 1600 | 35.3 | 4.3 :grin: | 8.2

The code is in https://github.com/rhattersley/iris/tree/decoupler.